Phase XXIV: Fair Launch & Integrity Protocol
FRONTEND DATA INTEGRITY:
- Inject TOTAL_SUPPLY_QNTO=21,000,000,000 constant across Codex.tsx, TokenSale.tsx, TelemetryDashboard.tsx
- Expand Codex.tsx tokenomics grid to 5 cards: Total Supply, Community 80%, Eco/Dev 15%, Liquidity 5%, Fair Launch ✓
- Add SAGA AI micro-fee callout documenting 0.000001 QNTO base fee and 8M TPS subsidy threshold
- Implement mathematical overflow guard in TokenSale.tsx (clamps to MAX_PURCHASABLE_QNTO ceiling)
- Add Supply Cap indicator card to TelemetryDashboard.tsx (21B QNTO hard cap)

DYNAMIC MICRO-FEE ENGINE:
- Adjust SAGA_SUBSIDY_THRESHOLD from 0.9 to 0.8 (80% of 10M TPS = 8M TPS) in gas_fee_model.rs
- Enhanced doc comments for anti-spam threshold behavior

FRICTIONLESS MINING CLI:
- Add ExportKey and ShowMnemonic subcommands to qanto.rs CLI
- Add --export-key and --show-mnemonic flags to start_node.rs
- Password prompts strictly isolated to these sensitive key operations only
- Node boot remains completely frictionless (WALLET_PASSWORD env var, defaults to empty)
- Add export_private_key_hex() and get_mnemonic() methods to wallet.rs

CODEBASE POLISH:
- cargo clippy --fix applied across workspace
- 19/19 tests passed, 0 failed, 0 ignored (100% pass rate)

+++ Protocol_Core/qanto-node/src/config.rs
@@ -981,7 +981,7 @@ impl Config {
 
         // Validate P2P heartbeat interval consistency
         // Heartbeat should be at most 2x target block time
-        if self.p2p.heartbeat_interval as u64 > self.target_block_time * 2 {
+        if self.p2p.heartbeat_interval > self.target_block_time * 2 {
             return Err(ConfigError::Validation(format!(
                 "p2p.heartbeat_interval ({}ms) is too large relative to target_block_time ({}ms). \
                  Heartbeat interval should be at most 2x target block time ({}ms)",
